+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Add tinyMCE plugin that adds helper class to elements (T-17631)
 * Use eager loading instead of lazy if `native_lazyload_tag` is used in the first block (T-17629)
+* Change to minified version of instant.page 5.2.0 (T-23371)
 
 ### 3.1.1: 2024-12-02